fix: isolate and abstract visitors & requests
