+++ src/modules/cp/components/ApproveRequestBlade/ApproveRequestBlade.utils.ts
@@ -1,7 +1,4 @@
-import { format } from 'date-fns';
-import { i18n } from 'next-i18next';
-
-import { getLocaleFromi18nLanguage } from '@shared/utils';
+import { formatWithLocale } from '@shared/utils';
 
 export const formatApproveRequestAccessDate = (date?: Date): string => {
 	return formatWithLocale('P', date);
@@ -9,9 +6,4 @@ export const formatApproveRequestAccessDate = (date?: Date): string => {
 
 export const formatApproveRequestAccessTime = (date?: Date): string => {
 	return formatWithLocale('p', date);
-};
-
-export const formatWithLocale = (formatString: string, date?: Date): string => {
-	const locale = getLocaleFromi18nLanguage(i18n?.language || '');
-	return date ? format(date, formatString, { locale }) : '';
 };
