+++ src/modules/cp/const/requests.const.tsx
@@ -1,19 +1,14 @@
 import { TabProps } from '@meemoo/react-components';
-import clsx from 'clsx';
-import { format, isWithinInterval } from 'date-fns';
 import { i18n, TFunction } from 'next-i18next';
 import { Column } from 'react-table';
 import { NumberParam, StringParam, withDefault } from 'use-query-params';
 
 import { RequestStatusBadge } from '@cp/components';
 import { RequestStatusAll } from '@cp/types';
-import { Icon, UnreadMarker } from '@shared/components';
+import { Icon } from '@shared/components';
 import { SEARCH_QUERY_KEY } from '@shared/const';
 import { SortDirectionParam } from '@shared/helpers';
 import { VisitInfo, VisitInfoRow, VisitStatus } from '@shared/types';
-import { VisitTimeframe } from '@visits/types';
-
-import styles from './requests.module.scss';
 
 export const RequestTablePageSize = 20;
 
@@ -59,7 +54,7 @@ export const RequestTableColumns = (
 		Cell: ({ row }: VisitInfoRow) => {
 			return (
 				<a
-					className="u-color-neutral p-cp-requests__link"
+					className="u-color-neutral c-table__link"
 					href={`mailto:${row.original.visitorMail}`}
 					onClick={(e) => e.stopPropagation()}
 				>
@@ -91,87 +86,6 @@ export const RequestTableColumns = (
 		id: 'cp-requests-table-actions',
 		Cell: () => {
 			return <Icon className="p-cp-requests__actions" name="dots-vertical" />;
-		},
-	},
-];
-
-export const CP_ADMIN_VISITS_QUERY_PARAM_CONFIG = {
-	timeframe: withDefault(StringParam, RequestStatusAll.ALL),
-	[SEARCH_QUERY_KEY]: withDefault(StringParam, undefined),
-	page: withDefault(NumberParam, 1),
-	orderProp: withDefault(StringParam, undefined),
-	orderDirection: withDefault(SortDirectionParam, undefined),
-};
-
-export const visitsStatusFilters = (): TabProps[] => {
-	return [
-		{
-			id: RequestStatusAll.ALL,
-			label: i18n?.t('modules/cp/const/requests___alle'),
-		},
-		{
-			id: VisitTimeframe.ACTIVE,
-			label: i18n?.t('modules/cp/const/requests___actief'),
-		},
-		{
-			id: VisitTimeframe.PAST,
-			label: i18n?.t('modules/cp/const/requests___historiek'),
-		},
-	];
-};
-
-export const VisitsTableColumns = (
-	i18n: { t: TFunction } = { t: (x: string) => x }
-): Column<VisitInfo>[] => [
-	{
-		Header: i18n.t('modules/cp/const/requests___naam') || '',
-		accessor: 'visitorName',
-	},
-	{
-		Header: i18n.t('modules/cp/const/requests___emailadres') || '',
-		accessor: 'visitorMail',
-		Cell: ({ row }: VisitInfoRow) => {
-			return (
-				<a
-					className="u-color-neutral p-cp-visitors__link"
-					href={`mailto:${row.original.visitorMail}`}
-					onClick={(e) => e.stopPropagation()}
-				>
-					{row.original.visitorMail}
-				</a>
-			);
-		},
-	},
-	{
-		Header: i18n.t('modules/cp/const/requests___toegang') || '',
-		accessor: 'startAt',
-		Cell: ({ row }: VisitInfoRow) => {
-			const active = isWithinInterval(new Date(), {
-				start: new Date(row.original.startAt || ''),
-				end: new Date(row.original.endAt || ''),
-			});
-			return (
-				<span className={clsx('u-color-neutral', styles['p-visits__access'])}>
-					<UnreadMarker active={active} />
-					{format(new Date(row.original.startAt || ''), 'dd/MM/yyyy HH:mm') +
-						' - ' +
-						format(new Date(row.original.endAt || ''), 'dd/MM/yyyy HH:mm')}
-				</span>
-			);
-		},
-	},
-	{
-		Header: i18n.t('modules/cp/const/requests___goedgekeurd-door') || '',
-		accessor: 'status',
-		Cell: () => {
-			return <span className="u-color-neutral">Noah Peeters</span>; // TODO Add column to database to track updatedBy
-		},
-	},
-	{
-		Header: '',
-		id: 'cp-visitors-histories-table-actions',
-		Cell: () => {
-			return <Icon className="p-cp-visitors-histories__actions" name="dots-vertical" />;
 		},
 	},
 ];
