--- src/modules/cp/const/requests.module.scss
@@ -1,5 +1,0 @@
-.p-visits__access {
-  display: flex;
-  align-items: center;
-}
-
