+++ src/modules/shared/utils/index.ts
@@ -1,5 +1,6 @@
+export * from './create-url';
+export * from './dates';
 export * from './i18n';
 export * from './is-browser';
 export * from './seo';
 export * from './tos';
-export * from './create-url';
