+++ src/styles/pages/cp/_visitors.scss
@@ -33,16 +33,9 @@ $page: "p-cp-visitors";
 		transform: translateY(0.2rem); // shift if to center on single-line
 	}
 
-    &__link {
-		text-decoration: none;
-		transition: color $animate-default;
-
-		&:hover,
-		&:active,
-		&:focus,
-		&:focus-within {
-			color: $teal !important;
-		}
+	&__access {
+		display: flex;
+		align-items: center;
 	}
 
 	tr {
